requirements.txt
pandas
numpy
matplotlib
requests
bitfinex-tencars